buggs de redirection regles,test du user ok, correction de la mise en forme

+++ src/Controller/QuestionController.php
@@ -43,6 +43,16 @@ class QuestionController extends AbstractController
     #[Route('/user/{username}/{filter}', name: 'user.index', requirements: ['username' => '^[a-z0-9_-]{4,15}$', 'filter' => 'all|service_[0-9]+'])]
     public function userIndex(EntityManagerInterface $entityManager, string $username, $filter = 'all'): Response
     {
+        if($this->getUser()){
+            $roles = $this->getUser()->getRoles();
+            if(in_array('ROLE_USER',$roles))
+            {
+                if(in_array('ROLE_ADMIN',$roles)) {
+                    return $this->redirectToRoute('blog.question.admin.index',['username' => $this->getUser()->getUserIdentifier(),'filter'=>$filter]);
+                }
+                //return $this->redirectToRoute('blog.question.user.index',['username' => $this->getUser()->getUserIdentifier(),'filter'=>$filter]);
+            }
+        }
         $services = $entityManager->getRepository(Service::class)->findAll();
 
         if ($filter === 'all') {
@@ -108,6 +118,16 @@ class QuestionController extends AbstractController
     #[Route('/{id}/show/user/{username}', name: 'user.show', requirements: ['username' => '^[a-z0-9_-]{4,15}$', 'id' => Requirement::DIGITS])]
     public function userShow(EntityManagerInterface $entityManager, $id, $username): Response
     {
+        if($this->getUser()){
+            $roles = $this->getUser()->getRoles();
+            if(in_array('ROLE_USER',$roles))
+            {
+                if(in_array('ROLE_ADMIN',$roles)) {
+                    return $this->redirectToRoute('blog.question.admin.show',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+                }
+               // return $this->redirectToRoute('blog.question.user.show',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+            }
+        }
         $question = $entityManager->getRepository(Question::class)->find($id);
 
         if (!$question) {
@@ -122,19 +142,18 @@ class QuestionController extends AbstractController
             'username' => $username,
         ]);
     }
-    #[IsGranted('ROLE_USER')]
+    #[IsGranted('ROLE_ADMIN')]
     #[Route('/{id}/show/admin/{username}', name: 'admin.show', requirements: ['username' => '^[a-z0-9_-]{4,15}$', 'id' => Requirement::DIGITS])]
     public function adminShow(EntityManagerInterface $entityManager, $id, $username): Response
     {
-        $question = $entityManager->getRepository(Question::class)->find($id);
 
+        $question = $entityManager->getRepository(Question::class)->find($id);
         if (!$question) {
             $this->addFlash('error', "Cette question n'existe pas");
             return $this->redirectToRoute('blog.question.admin.index', [
                 'username' => $username,
             ]);
         }
-
         return $this->render('question/admin/show.html.twig', [
             'question' => $question,
             'username' => $username,
@@ -150,7 +169,17 @@ class QuestionController extends AbstractController
     #[IsGranted('ROLE_USER')]
     #[Route('/ask/user/{username}', name: 'user.create', requirements: ['username' => '^[a-z0-9_-]{4,15}$'])]
     public function userCreate(Request $request, EntityManagerInterface $entityManager, $username): Response
-    {
+        {
+            if($this->getUser()){
+            $roles = $this->getUser()->getRoles();
+            if(in_array('ROLE_USER',$roles))
+            {
+                if(in_array('ROLE_ADMIN',$roles)) {
+                    return $this->redirectToRoute('blog.question.admin.create',['username' => $this->getUser()->getUserIdentifier()]);
+                }
+               // return $this->redirectToRoute('blog.question.user.create',['username' => $this->getUser()->getUserIdentifier()]);
+            }
+        }
         if (!$this->isGranted('ROLE_USER')) {
             $this->addFlash("alert", "Connectez-vous ou inscrivez-vous pour créer une question");
             return $this->redirectToRoute('app_login');
@@ -183,6 +212,7 @@ class QuestionController extends AbstractController
     #[Route('/ask/admin/{username}', name: 'admin.create', requirements: ['username' => '^[a-z0-9_-]{4,15}$'])]
     public function adminCreate(Request $request, EntityManagerInterface $entityManager, string $username): Response
     {
+
         $question = new Question();
         $form = $this->createForm(QuestionType::class, $question);
         $form->handleRequest($request);
@@ -211,21 +241,27 @@ class QuestionController extends AbstractController
     #[Route('/{id}/edit', name: 'default.edit', requirements: ['id' => Requirement::DIGITS])]
     public function edit(Request $request, EntityManagerInterface $entityManager, $id): Response
     {
-        $question = $entityManager->getRepository(Question::class)->find($id);
 
+        if($this->getUser()){
+            $roles = $this->getUser()->getRoles();
+            if(in_array('ROLE_USER',$roles))
+            {
+                if(in_array('ROLE_ADMIN',$roles)) {
+                    return $this->redirectToRoute('blog.question.admin.edit',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+                }
+                return $this->redirectToRoute('blog.question.user.edit',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+            }
+        }
+        $question = $entityManager->getRepository(Question::class)->find($id);
         if (!$question) {
             $this->addFlash('error', "Cette question n'existe pas");
             return $this->redirectToRoute('blog.question.default.index');
         }
-
         $form = $this->createForm(QuestionType::class, $question);
-
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
-
             $this->addFlash('success', 'Question modifiée avec succès');
-
             return $this->redirectToRoute('blog.question.default.index');
         }
 
@@ -239,6 +275,17 @@ class QuestionController extends AbstractController
     {
         $question = $entityManager->getRepository(Question::class)->find($id);
 
+        if($this->getUser()){
+            $roles = $this->getUser()->getRoles();
+            if(in_array('ROLE_USER',$roles))
+            {
+                if(in_array('ROLE_ADMIN',$roles)) {
+                    return $this->redirectToRoute('blog.question.admin.edit',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+                }
+              //  return $this->redirectToRoute('blog.question.user.create',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+            }
+        }
+
         if (!$question) {
             $this->addFlash('error', "Cette question n'existe pas");
             return $this->redirectToRoute('blog.question.user.index', ['username' => $username]);
@@ -249,12 +296,9 @@ class QuestionController extends AbstractController
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
-
             $this->addFlash('success', 'Question modifiée avec succès');
-
             return $this->redirectToRoute('blog.question.user.index', ['username' => $username]);
         }
-
         return $this->render('question/user/edit.html.twig', [
             'form' => $form,
             'username' => $username,
@@ -265,23 +309,17 @@ class QuestionController extends AbstractController
     public function adminEdit(Request $request, EntityManagerInterface $entityManager, $id, $username): Response
     {
         $question = $entityManager->getRepository(Question::class)->find($id);
-
         if (!$question) {
             $this->addFlash('error', "Cette question n'existe pas");
             return $this->redirectToRoute('blog.question.admin.index', ['username' => $username]);
         }
-
         $form = $this->createForm(QuestionType::class, $question);
-
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
-
             $this->addFlash('success', 'Question modifiée avec succès');
-
             return $this->redirectToRoute('blog.question.admin.index', ['username' => $username]);
         }
-
         return $this->render('question/admin/edit.html.twig', [
             'form' => $form,
             'username' => $username,
@@ -292,36 +330,48 @@ class QuestionController extends AbstractController
     #[Route('/{id}/delete', name: 'default.delete', requirements: ['id' => Requirement::DIGITS])]
     public function delete(EntityManagerInterface $entityManager, $id): Response
     {
+        if($this->getUser()){
+            $roles = $this->getUser()->getRoles();
+            if(in_array('ROLE_USER',$roles))
+            {
+                if(in_array('ROLE_ADMIN',$roles)) {
+                    return $this->redirectToRoute('blog.question.admin.delete',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+                }
+                  return $this->redirectToRoute('blog.question.user.delete',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+            }
+        }
         $question = $entityManager->getRepository(Question::class)->find($id);
-
         if (!$question) {
             $this->addFlash('error', "Cette question n'existe pas");
             return $this->redirectToRoute('blog.question.default.index');
         }
-
         $entityManager->remove($question);
         $entityManager->flush();
-
         $this->addFlash('success', 'Question supprimée avec succès');
-
         return $this->redirectToRoute('blog.question.default.index');
     }
     #[IsGranted('ROLE_USER')]
     #[Route('/{id}/delete/user/{username}', name: 'user.delete', requirements: ['username' => '^[a-z0-9_-]{4,15}$', 'id' => Requirement::DIGITS])]
     public function userDelete(EntityManagerInterface $entityManager, $id, $username): Response
     {
+        if($this->getUser()){
+            $roles = $this->getUser()->getRoles();
+            if(in_array('ROLE_USER',$roles))
+            {
+                if(in_array('ROLE_ADMIN',$roles)) {
+                    return $this->redirectToRoute('blog.question.admin.delete',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+                }
+//                return $this->redirectToRoute('blog.question.user.delete',['username' => $this->getUser()->getUserIdentifier(),'id'=>$id]);
+            }
+        }
         $question = $entityManager->getRepository(Question::class)->find($id);
-
         if (!$question) {
             $this->addFlash('error', "Cette question n'existe pas");
             return $this->redirectToRoute('blog.question.user.index', ['username' => $username]);
         }
-
         $entityManager->remove($question);
         $entityManager->flush();
-
         $this->addFlash('success', 'Question supprimée avec succès');
-
         return $this->redirectToRoute('blog.question.user.index', ['username' => $username]);
     }
     #[IsGranted('ROLE_ADMIN')]
